+++ src/data/store/pool/mod.rs
@@ -1,11 +1,14 @@
-// devela/src/data/layout/pool/mod.rs
+// devela/src/data/store/pool/mod.rs
 //
-//!
+//! Reusable stores with stable handles and individual reclamation.
 //
+
+mod define; // pool!
 
 crate::structural_mods! { // _mods
     _mods {
-        // pub use super::{
-        // };
+        pub use super::{
+            define::*,
+        };
     }
 }